+++ src/main/java/com/DietasYRutinasOnline/repository/NotificacionRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.DietasYRutinasOnline.entity.Notificacion;
-import com.DietasYRutinasOnline.entity.TipoUsuario;
+import com.DietasYRutinasOnline.entity.Rol;
 import com.DietasYRutinasOnline.entity.Transaccion;
 import com.DietasYRutinasOnline.entity.Usuario;
 
 @Repository
-public interface NotificacionRepository extends JpaRepository<Notificacion, Integer>{
+public interface NotificacionRepository extends JpaRepository<Notificacion, Long>{
 
 	//List<Notificacion> findByEstado(String estado);
 	List<Notificacion> findByRol(String rol);
